backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
asyncpg==0.31.0
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
debugpy==1.8.19
exceptiongroup==1.3.1
fastapi==0.124.4
graphql-core==3.2.7
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
lia-web==0.2.3
packaging==25.0
passlib==1.7.4
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
strawberry-graphql==0.287.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
